Implementado BFS

+++ ClusteringHumano/src/modelo/Grafo.java
@@ -1,7 +1,9 @@
 package modelo;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 public class Grafo<T> {
@@ -20,20 +22,46 @@ public class Grafo<T> {
 	}
 
 	public void agregarArista(Vertice<T> verticeA, Vertice<T> verticeB) {
+		crearArista(verticeA, verticeB, 0);
+	}
+
+	public void agregarArista(Vertice<T> verticeA, Vertice<T> verticeB, int cargaArista) {
+		crearArista(verticeA, verticeB, cargaArista);
+	}
+
+	private void crearArista(Vertice<T> verticeA, Vertice<T> verticeB, int cargaArista) {
+		if (!existeVertice(verticeA))
+			throw new IllegalArgumentException("El vértice A indicado no pertenece al grafo.");
+		if (!existeVertice(verticeB))
+			throw new IllegalArgumentException("El vértice B indicado no pertenece al grafo.");
+
+		if (sonElMismoVertice(verticeA, verticeB))
+			throw new IllegalArgumentException("Los vértices indicados son el mismo vértice. No se admiten ciclos.");
+
 		HashSet<Vertice<T>> listaVecinosA = vertices.get(verticeA);
 		HashSet<Vertice<T>> listaVecinosB = vertices.get(verticeB);
 
 		listaVecinosA.add(verticeB);
 		listaVecinosB.add(verticeA);
+
+//		Arista<T> nuevaArista = new Arista<>(verticeA, verticeB, cargaArista);
+//		aristas.add(nuevaArista);
 	}
 
-	public Set<Vertice<T>> getVecinos(Vertice<T> vertice) {
+	public boolean sonElMismoVertice(Vertice<T> verticeA, Vertice<T> verticeB) {
+		return verticeA.equals(verticeB);
+	}
+
+	public Set<Vertice<T>> obtenerVecinos(Vertice<T> vertice) {
+		if (!existeVertice(vertice))
+			throw new IllegalArgumentException("El vértice indicado no pertenece al grafo.");
+
 		HashSet<Vertice<T>> listaVecinos = vertices.get(vertice);
 
 		return listaVecinos;
 	}
 
-	public Set<Arista<T>> getAristas() {
+	public Set<Arista<T>> obtenerAristas() {
 		return aristas;
 	}
 
@@ -41,21 +69,56 @@ public class Grafo<T> {
 		return vertices.size();
 	}
 
-	public int getCantidadVecinos(Vertice<T> vertice) {
-		return getVecinos(vertice).size();
+	public boolean estaVacio() {
+		return tamanio() == 0;
+	}
+
+	public boolean noTieneVecinos(Vertice<T> vertice) {
+		return obtenerCantidadVecinos(vertice) == 0;
+	}
+
+	public int obtenerCantidadVecinos(Vertice<T> vertice) {
+		if (!existeVertice(vertice))
+			throw new IllegalArgumentException("El vértice indicado no pertenece al grafo.");
+
+		return obtenerVecinos(vertice).size();
 	}
 
 	public void eliminarVertice(Vertice<T> vertice) {
-		vertices.remove(vertice);
+		if (existeVertice(vertice)) {
+			vertices.remove(vertice);
 
-		eliminarVerticeDeListasDeVecinos(vertice);
+			eliminarVerticeDeListasDeVecinos(vertice);
+		}
 	}
 
 	private void eliminarVerticeDeListasDeVecinos(Vertice<T> verticeAEliminar) {
 		Set<Vertice<T>> conjuntoVertices = vertices.keySet();
 
 		for (Vertice<T> verticeActual : conjuntoVertices) {
-			getVecinos(verticeActual).remove(verticeAEliminar);
+			obtenerVecinos(verticeActual).remove(verticeAEliminar);
 		}
+	}
+
+	public boolean sonVecinos(Vertice<T> verticeA, Vertice<T> verticeB) {
+		if (!existeVertice(verticeA) || !existeVertice(verticeB))
+			return false;
+
+		boolean BEsVecinoDeA = obtenerVecinos(verticeA).contains(verticeB);
+		boolean AEsVecinoDeB = obtenerVecinos(verticeB).contains(verticeA);
+
+		return BEsVecinoDeA && AEsVecinoDeB;
+	}
+
+	private boolean existeVertice(Vertice<T> vertice) {
+		return vertices.containsKey(vertice);
+	}
+
+//	private boolean existeArista(Arista<T> arista) {
+//		return aristas.contains(arista);
+//	}
+
+	public List<Vertice<T>> obtenerVertices() {
+		return new ArrayList<Vertice<T>>(vertices.keySet());
 	}
 }
